Refactoring the CLI project to be reused across offline and onchain scenarios
