+++ source/Library.sln
@@ -7,6 +7,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Library", "library\Library.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Library.Test", "library-test\Library.Test.csproj", "{7277F72D-FFDC-4646-B6F5-B1E750470AD3}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cli", "cli\Cli.csproj", "{396367C7-A023-4676-8B61-0B83D283EADD}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -24,5 +26,9 @@ Global
 		{7277F72D-FFDC-4646-B6F5-B1E750470AD3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7277F72D-FFDC-4646-B6F5-B1E750470AD3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7277F72D-FFDC-4646-B6F5-B1E750470AD3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{396367C7-A023-4676-8B61-0B83D283EADD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{396367C7-A023-4676-8B61-0B83D283EADD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{396367C7-A023-4676-8B61-0B83D283EADD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{396367C7-A023-4676-8B61-0B83D283EADD}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
