+++ source/offline-poc/offline-cli-test/OfflineCli.Test.csproj
@@ -29,8 +29,7 @@
   <ItemGroup>
     <ProjectReference Include="..\..\library-test\Library.Test.csproj" />
     <ProjectReference Include="..\..\library\Library.csproj" />
-    <ProjectReference Include="..\..\library-test\Library.Test.csproj" />
-    <ProjectReference Include="..\..\offline-poc\offline-client\OfflineClient.csproj" />
+    <ProjectReference Include="..\..\cli\Cli.csproj" />
   </ItemGroup>
 
 </Project>
